Consolidated snippet and search snippet list
